organized vision and set constants to comp bot
